+++ Studyroom/src/main/resources/context/Context_2_myBatis.java
@@ -28,5 +28,10 @@ public class Context_2_myBatis {
 		
 		return factoryBean.getObject();
 	}
+	
+	@Bean
+	public SqlSessionTemplate sqlSessionTemplate(SqlSessionFactory factoryBean) {
+		return new SqlSessionTemplate(factoryBean);
+   }
 
 }
